Implement a fallback TranslationProvider mechanism
Until now, the original provider (the one for the target language most of the time) may not be able provide the requested message, in those cases we returned the message key right away.

But now we will also try to get the message from the configured default language, if unsuccessful, we ask the fallback provider, which is usually getting messages from the en_US bundled messages file. And only if the fallback provider does not return a message, then we return the message key.