Introduce NegativityAccountManager
It is used as a step between user code and a NegativityAccountStorage, made to not save accounts on the proxy server.
This way we do not overwrite newer data (from the game server), but still give the proxy access to accounts, specifically to get a player's language

Right now game servers do not notify companion plugins when an account has been updated. Players need to reconnect to the proxy in order to refresh their account data. This will be fixed later.

Adapter#getNegativityAccount and Adapter#invalidateAccount have been deprecated. Their usages will be replaced in a future commit. We may want to keep the methods deprecated for some time to not break backwards compatibility.