Update Gradle and ForgeGradle
We now use ForgeGradle 4 for the sponge plugin.
This version is future-proof and shouldn't break as much as previous versions of the plugin.
Unfortunately it does not work on Gradle 7, thankfully FG5 will fix this.

A downside of using FG4 is that we can't filter out client classes (or I haven't found how yet.)

Some repositories do not include POMs and this became mandatory by default in Gradle 6, hence why artifact() is needed on many repos (it relaxes this restriction.)